+++ aqua/util/output_saver.py
@@ -1,5 +1,3 @@
-"""Class for saving outputs, including NetCDF, PDF, and PNG files"""
-
 from aqua.logger import log_configure, log_history
 from aqua.util import add_pdf_metadata, add_png_metadata, update_metadata_with_date, ConfigPath
 import os
@@ -53,7 +51,7 @@ class OutputSaver:
 
     def generate_name(self, diagnostic_product: str = None, var: str = None, model_2: str = None, exp_2: str = None,
                       time_start: str = None, time_end: str = None, time_precision: str = 'ymd', area: str = None,
-                      suffix: str = 'nc', catalog: str = None, catalog_2: str = None, **kwargs) -> str:
+                      suffix: str = 'nc', catalog_2: str = None, **kwargs) -> str:
         """
         Generate a filename based on provided parameters and additional user-defined keywords,
         including precise time intervals.
@@ -68,7 +66,6 @@ class OutputSaver:
             time_precision (str, optional): Precision for time representation ('y', 'ym', 'ymd', 'ymdh', etc.).
             area (str, optional): The geographical area covered by the data.
             suffix (str, optional): The file extension/suffix indicating file type.
-            catalog (str, optional): Catalog where to search for the triplet. Default to None will allow for autosearch in the installed catalogs.
             catalog_2 (str, optional): The second catalog, for comparative studies. Default to None will allow for autosearch in the installed catalogs.
             **kwargs: Arbitrary keyword arguments provided by the user for additional customization.
 
@@ -107,7 +104,7 @@ class OutputSaver:
         additional_parts = [f"{key}_{value}" for key, value in sorted(kwargs.items())]
 
         parts = [part for part in [self.diagnostic, self.diagnostic_product, var,
-                                   self.model, self.exp, catalog if catalog else self.catalog, model_2, exp_2, catalog_2, area] if part]
+                                   self.model, self.exp, self.catalog, model_2, exp_2, catalog_2, area] if part]
         parts.extend(time_parts)
         parts.extend(additional_parts)
         parts.append(suffix)
@@ -118,7 +115,7 @@ class OutputSaver:
 
     def save_netcdf(self, dataset: xr.Dataset, path: str = None, diagnostic_product: str = None, var: str = None,
                     model_2: str = None, exp_2: str = None, time_start: str = None, time_end: str = None,
-                    time_precision: str = 'ymd', area: str = None, metadata: dict = None, catalog: str = None, catalog_2: str = None, **kwargs) -> str:
+                    time_precision: str = 'ymd', area: str = None, metadata: dict = None, catalog_2: str = None, **kwargs) -> str:
         """
         Save a netCDF file with a dataset to a specified path, with support for additional filename keywords and
         precise time intervals.
@@ -135,7 +132,6 @@ class OutputSaver:
             time_precision (str, optional): Precision for time representation ('y', 'ym', 'ymd', 'ymdh', etc.).
             area (str, optional): The geographical area covered by the data.
             metadata (dict, optional): Additional metadata to include in the netCDF file.
-            catalog (str, optional): Catalog where to search for the triplet. Default to None will allow for autosearch in the installed catalogs.
             catalog_2 (str, optional): The second catalog, for comparative studies. Default to None will allow for autosearch in the installed catalogs.
             **kwargs: Additional keyword arguments for more flexible filename customization.
 
@@ -144,7 +140,7 @@ class OutputSaver:
         """
         filename = self.generate_name(diagnostic_product=diagnostic_product, var=var,
                                       model_2=model_2, exp_2=exp_2, time_start=time_start, time_end=time_end,
-                                      time_precision=time_precision, area=area, suffix='nc', catalog=catalog, catalog_2=catalog_2, **kwargs)
+                                      time_precision=time_precision, area=area, suffix='nc', catalog_2=catalog_2, **kwargs)
 
         if path is None:
             path = self.default_path
@@ -174,7 +170,7 @@ class OutputSaver:
 
     def save_pdf(self, fig: Figure, path: str = None, diagnostic_product: str = None, var: str = None,
                  model_2: str = None, exp_2: str = None, time_start: str = None, time_end: str = None,
-                 time_precision: str = 'ymd', area: str = None, metadata: dict = None, dpi: int = 300, catalog: str = None, catalog_2: str = None, **kwargs) -> str:
+                 time_precision: str = 'ymd', area: str = None, metadata: dict = None, dpi: int = 300, catalog_2: str = None, **kwargs) -> str:
         """
         Save a PDF file with a matplotlib figure to the provided path, with support for additional filename keywords and
         precise time intervals.
@@ -192,7 +188,6 @@ class OutputSaver:
             area (str, optional): The geographical area covered by the data.
             metadata (dict, optional): Additional metadata to include in the PDF file.
             dpi (int, optional): The resolution of the saved PDF file. Default is 300.
-            catalog (str, optional): Catalog where to search for the triplet. Default to None will allow for autosearch in the installed catalogs.
             catalog_2 (str, optional): The second catalog, for comparative studies. Default to None will allow for autosearch in the installed catalogs.
             **kwargs: Additional keyword arguments for more flexible filename customization.
 
@@ -206,7 +201,7 @@ class OutputSaver:
             path = self.default_path
         filename = self.generate_name(diagnostic_product=diagnostic_product, var=var, model_2=model_2, exp_2=exp_2,
                                       time_start=time_start, time_end=time_end, time_precision=time_precision, area=area,
-                                      suffix='pdf', catalog=catalog, catalog_2=catalog_2, **kwargs)
+                                      suffix='pdf', catalog_2=catalog_2, **kwargs)
         full_path = os.path.join(path, filename)
 
         if not self.rebuild and os.path.exists(full_path):
@@ -231,7 +226,7 @@ class OutputSaver:
 
     def save_png(self, fig: Figure, path: str = None, diagnostic_product: str = None, var: str = None,
                  model_2: str = None, exp_2: str = None, time_start: str = None, time_end: str = None,
-                 time_precision: str = 'ymd', area: str = None, metadata: dict = None, dpi: int = 300, catalog: str = None, catalog_2: str = None, **kwargs) -> str:
+                 time_precision: str = 'ymd', area: str = None, metadata: dict = None, dpi: int = 300, catalog_2: str = None, **kwargs) -> str:
         """
         Save a PNG file with a matplotlib figure to a provided path, with support for additional filename keywords and
         precise time intervals.
@@ -249,7 +244,6 @@ class OutputSaver:
             area (str, optional): The geographical area covered by the data.
             metadata (dict, optional): Additional metadata to include in the PNG file.
             dpi (int, optional): The resolution of the saved PNG file. Default is 300.
-            catalog (str, optional): Catalog where to search for the triplet. Default to None will allow for autosearch in the installed catalogs.
             catalog_2 (str, optional): The second catalog, for comparative studies. Default to None will allow for autosearch in the installed catalogs.
             **kwargs: Additional keyword arguments for more flexible filename customization.
 
@@ -261,7 +255,7 @@ class OutputSaver:
         """
         filename = self.generate_name(diagnostic_product=diagnostic_product, var=var, model_2=model_2, exp_2=exp_2,
                                       time_start=time_start, time_end=time_end, time_precision=time_precision,
-                                      area=area, suffix='png', catalog=catalog, catalog_2=catalog_2, **kwargs)
+                                      area=area, suffix='png', catalog_2=catalog_2, **kwargs)
 
         if path is None:
             path = self.default_path